+++ src/App.js
@@ -14,6 +14,8 @@ import { SocketProvider } from "./context/SocketContext";
 import BookingDetails from "./pages/user/BookingDetails";
 import ErrorPage from "./components/ErrorPage";
 import Auth from "./pages/Auth";
+import { GoogleOAuthProvider } from "@react-oauth/google";
+import EmailVerification from "./pages/EmailVerification";
 
 function App() {
   return (
@@ -50,7 +52,11 @@ export const appRouter = createBrowserRouter([
       <NetworkErrorProvider>
         <ErrorBoundary>
           <SocketProvider>
-            <App />
+            <GoogleOAuthProvider
+              clientId={process.env.REACT_APP_GOOGLE_CLIENT_ID}
+            >
+              <App />
+            </GoogleOAuthProvider>
           </SocketProvider>
         </ErrorBoundary>
       </NetworkErrorProvider>
@@ -84,6 +90,10 @@ export const appRouter = createBrowserRouter([
       {
         path: "/order/:id",
         element: <BookingDetails />,
+      },
+      {
+        path: "/auth/verify-email",
+        element: <EmailVerification />,
       },
     ],
   },