+++ CHANGELOG.md
@@ -1,9 +1,22 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## [Unreleased]
+### [1.2.0]
 ### Added
-- Changed login and sign up page ui.
+- Social login feature implemented (Google)
+- OTP based registration system implemented
+- OTP rate limiting added prevent brute-force attack.
+
+### Fixed
+- Auth form value visible bug fixed
+
+## [1.1.1] 27/09/25
+### Fixed
+- Sign in button not properly visible in small device issue fixed.
+
+## [1.1.0] 27/09/25
+### Added
+- New auth page added for authentication.
 
 ## [1.0.0] - 27/09/25
 ### Added